Adds topic_prefix and topic_suffix options.
When present topic_prefix and topic_suffix preppends or append its value to
all the topic names were messages are emitted, both for the log aggregation topic
and the topics for emit calls.

This feature is useful to namespace topics if you have a single kafka cluster storing
messages for more than one environment.